Rewrite WDL's translation of step node (big change)
Golly gosh this was a hard one, but I got it. Condensed the scattering and secondary logic into a shorter,
much more succinct block that still passes all the tests (so far).

Really, it was ensuring that the values are unwrapped correctly (essentially replace dotted_source with
`unwrap_expression` (as the input / stepoutput selectors are handled there) and rely on earlier Janis
checks to prohibit expressions when secondary files are involved.